feat: 친구 검색 및 소켓 이벤트 처리 로직 분리 및 리팩토링
- useFriendSearch 훅 생성으로 검색 로직 분리 및 상태 관리 개선
- useFriendSocketEvents 훅 추가로 소켓 이벤트 처리 로직 분리
- FriendsProvider 리팩토링: 소켓 이벤트 연결 로직 간소화
- debounce 로직 및 searchQuery 상태 관리 개선
- RoomInvitationModal 동작 관련 로직 최적화
- FriendRow 스타일 및 UI 디테일 업데이트
