+++ frontend/src/features/friends/contexts/FriendsProvider.tsx
@@ -1,36 +1,33 @@
-import { PropsWithChildren, useCallback, useEffect, useState } from 'react';
+import { PropsWithChildren, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useLocation } from 'react-router-dom';
-import useModal from '@/components/@common/Modal/useModal';
-import useToast from '@/components/@common/Toast/useToast';
-import { useUserSocketSubscription } from '@/apis/websocket/hooks/useUserSocketSubscription';
 import { useUserSocket } from '@/apis/websocket/contexts/UserSocketContext';
 import { useAuth } from '@/features/auth/hooks/useAuth';
 import { tokenStore } from '@/features/auth/tokens';
 import { friendsApi } from '../api/friendsApi';
-import RoomInvitationModal from '../components/RoomInvitationModal';
-import {
-  Friend,
-  FriendPresenceEvent,
-  FriendRemovedEvent,
-  FriendRequestEvent,
-  FriendResponseEvent,
-  ReceivedRequest,
-  RoomInvitationEvent,
-  SentRequest,
-} from '../types';
+import { useFriendSocketEvents } from '../hooks/useFriendSocketEvents';
+import { Friend, ReceivedRequest, SentRequest } from '../types';
 import { FriendsContext, FriendsContextType } from './FriendsContext';
 
 export const FriendsProvider = ({ children }: PropsWithChildren) => {
   const { isAuthenticated } = useAuth();
-  const { showToast } = useToast();
-  const { openModal, closeModal } = useModal();
   const { reconnect } = useUserSocket();
   const { pathname } = useLocation();
-  const isInRoom = pathname.startsWith('/room/');
 
   const [friends, setFriends] = useState<Friend[]>([]);
   const [receivedRequests, setReceivedRequests] = useState<ReceivedRequest[]>([]);
   const [sentRequests, setSentRequests] = useState<SentRequest[]>([]);
+
+  // ref로 관리해 deps 없이 최신 값 참조
+  const isInRoomRef = useRef(false);
+  const reconnectRef = useRef(reconnect);
+
+  useEffect(() => {
+    isInRoomRef.current = pathname.startsWith('/room/');
+  }, [pathname]);
+
+  useEffect(() => {
+    reconnectRef.current = reconnect;
+  }, [reconnect]);
 
   // 로그인 시 초기 데이터 로드
   useEffect(() => {
@@ -43,9 +40,6 @@ export const FriendsProvider = ({ children }: PropsWithChildren) => {
 
     const load = async () => {
       try {
-        // 토큰 만료 감지: 백엔드는 만료 토큰으로도 STOMP 연결을 허용(sessionId fallback)하므로
-        // REST 호출 결과로만 실제 인증 상태를 확인할 수 있다.
-        // 갱신 전 토큰을 저장해 두고, 성공 후 비교해 갱신 여부를 판단한다.
         const tokenBefore = tokenStore.getAccessToken();
         const [friendsList, received, sent] = await Promise.all([
           friendsApi.getFriends(),
@@ -55,9 +49,8 @@ export const FriendsProvider = ({ children }: PropsWithChildren) => {
         setFriends(friendsList);
         setReceivedRequests(received);
         setSentRequests(sent);
-        // 토큰이 갱신됐으면 STOMP도 새 토큰으로 재연결
         if (tokenBefore !== tokenStore.getAccessToken()) {
-          reconnect();
+          reconnectRef.current();
         }
       } catch {
         // 초기 로드 실패 시 조용히 빈 상태 유지
@@ -65,104 +58,16 @@ export const FriendsProvider = ({ children }: PropsWithChildren) => {
     };
 
     load();
-  }, [isAuthenticated, reconnect]);
+  }, [isAuthenticated]);
 
-  // 친구 요청 수신
-  useUserSocketSubscription<FriendRequestEvent>(
-    '/user/queue/friends/requests',
-    useCallback(
-      (event) => {
-        const { data } = event;
-        const req: ReceivedRequest = {
-          requestId: data.requestId,
-          userId: data.fromUserId,
-          userCode: data.fromUserCode,
-          nickname: data.fromNickname,
-          createdAt: data.createdAt,
-        };
-        setReceivedRequests((prev) => [req, ...prev]);
-        showToast({
-          message: `${req.nickname ?? '누군가'} 님이 친구 요청을 보냈습니다`,
-          type: 'info',
-        });
-      },
-      [showToast]
-    ),
-    isAuthenticated
-  );
-
-  // 내가 보낸 요청 응답 수신
-  useUserSocketSubscription<FriendResponseEvent>(
-    '/user/queue/friends/responses',
-    useCallback(
-      (event) => {
-        const { data } = event;
-        setSentRequests((prev) => prev.filter((r) => r.requestId !== data.requestId));
-
-        if (data.accepted) {
-          const newFriend: Friend = {
-            userId: data.counterpartUserId,
-            userCode: data.counterpartUserCode,
-            nickname: data.counterpartNickname,
-            since: new Date().toISOString(),
-            online: false,
-          };
-          setFriends((prev) =>
-            prev.some((f) => f.userId === newFriend.userId) ? prev : [...prev, newFriend]
-          );
-          showToast({
-            message: `${data.counterpartNickname} 님과 친구가 되었습니다`,
-            type: 'success',
-          });
-        } else {
-          showToast({ message: '친구 요청이 거절되었습니다', type: 'info' });
-        }
-      },
-      [showToast]
-    ),
-    isAuthenticated
-  );
-
-  // 친구 끊기 알림 수신
-  useUserSocketSubscription<FriendRemovedEvent>(
-    '/user/queue/friends/removed',
-    useCallback((event) => {
-      setFriends((prev) => prev.filter((f) => f.userId !== event.data.removedByUserId));
-    }, []),
-    isAuthenticated
-  );
-
-  // 방 초대 수신
-  useUserSocketSubscription<RoomInvitationEvent>(
-    '/user/queue/rooms/invitations',
-    useCallback(
-      (event) => {
-        if (isInRoom) return;
-        const { data } = event;
-        openModal(
-          <RoomInvitationModal
-            inviterNickname={data.inviterNickname}
-            joinCode={data.joinCode}
-            onClose={closeModal}
-          />,
-          { title: '방 초대', showCloseButton: false }
-        );
-      },
-      [isInRoom, openModal, closeModal]
-    ),
-    isAuthenticated
-  );
-
-  // 친구 온라인/오프라인 전이
-  useUserSocketSubscription<FriendPresenceEvent>(
-    '/user/queue/friends/presence',
-    useCallback((event) => {
-      setFriends((prev) =>
-        prev.map((f) => (f.userId === event.userId ? { ...f, online: event.online } : f))
-      );
-    }, []),
-    isAuthenticated
-  );
+  // STOMP 이벤트 처리 — 별도 훅으로 분리
+  useFriendSocketEvents({
+    setFriends,
+    setReceivedRequests,
+    setSentRequests,
+    isAuthenticated,
+    isInRoomRef,
+  });
 
   const updateFriendOnline = useCallback((userId: number, online: boolean) => {
     setFriends((prev) => prev.map((f) => (f.userId === userId ? { ...f, online } : f)));
@@ -188,18 +93,31 @@ export const FriendsProvider = ({ children }: PropsWithChildren) => {
     setSentRequests((prev) => prev.filter((r) => r.requestId !== requestId));
   }, []);
 
-  const contextValue: FriendsContextType = {
-    friends,
-    receivedRequests,
-    sentRequests,
-    pendingReceivedCount: receivedRequests.length,
-    updateFriendOnline,
-    removeFriendFromStore,
-    addFriend,
-    addReceivedRequest,
-    removeReceivedRequest,
-    removeSentRequest,
-  };
+  const contextValue: FriendsContextType = useMemo(
+    () => ({
+      friends,
+      receivedRequests,
+      sentRequests,
+      pendingReceivedCount: receivedRequests.length,
+      updateFriendOnline,
+      removeFriendFromStore,
+      addFriend,
+      addReceivedRequest,
+      removeReceivedRequest,
+      removeSentRequest,
+    }),
+    [
+      friends,
+      receivedRequests,
+      sentRequests,
+      updateFriendOnline,
+      removeFriendFromStore,
+      addFriend,
+      addReceivedRequest,
+      removeReceivedRequest,
+      removeSentRequest,
+    ]
+  );
 
   return <FriendsContext.Provider value={contextValue}>{children}</FriendsContext.Provider>;
 };
