+++ frontend/src/features/home/components/FriendsTab/FriendsTab.tsx
@@ -1,10 +1,9 @@
 import styled from '@emotion/styled';
-import { ChangeEvent, useCallback, useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 import { useAuth } from '@/features/auth/hooks/useAuth';
-import { friendsApi } from '@/features/friends/api/friendsApi';
 import { useFriends } from '@/features/friends/hooks/useFriends';
-import { SearchedUser } from '@/features/friends/types';
 import useToast from '@/components/@common/Toast/useToast';
+import { useFriendSearch } from '@/features/home/hooks/useFriendSearch';
 import { theme } from '@/styles/theme';
 import FriendsList from './FriendsList';
 import LoginRequiredView from './LoginRequiredView';
@@ -15,12 +14,21 @@ import SentRequestsList from './SentRequestsList';
 type TabKey = '친구' | '받은 요청' | '보낸 요청';
 type SearchMode = '닉네임' | '유저코드';
 
-const DEBOUNCE_MS = 300;
-
 const FriendsTab = () => {
   const { isAuthenticated, user } = useAuth();
   const { friends, receivedRequests, sentRequests } = useFriends();
   const { showToast } = useToast();
+  const {
+    searchQuery,
+    searchMode,
+    searchResults,
+    searchLoading,
+    isSearching,
+    handleSearchChange,
+    handleModeChange,
+  } = useFriendSearch();
+
+  const [activeTab, setActiveTab] = useState<TabKey>('친구');
 
   const handleCopyCode = () => {
     if (!user) return;
@@ -28,66 +36,10 @@ const FriendsTab = () => {
     showToast({ message: '유저코드가 복사되었습니다', type: 'success' });
   };
 
-  const [activeTab, setActiveTab] = useState<TabKey>('친구');
-  const [searchQuery, setSearchQuery] = useState('');
-  const [searchMode, setSearchMode] = useState<SearchMode>('닉네임');
-  const [searchResults, setSearchResults] = useState<SearchedUser[]>([]);
-  const [searchLoading, setSearchLoading] = useState(false);
-  const debounceRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const searchIdRef = useRef(0);
-
-  const doSearch = useCallback(async (query: string, mode: SearchMode) => {
-    if (!query.trim()) {
-      setSearchResults([]);
-      return;
-    }
-    if (mode === '유저코드' && query.trim().length !== 5) return;
-
-    const id = ++searchIdRef.current;
-    try {
-      setSearchLoading(true);
-      const results =
-        mode === '닉네임'
-          ? await friendsApi.searchByNickname(query.trim())
-          : await friendsApi.searchByUserCode(query.trim());
-      if (id === searchIdRef.current) setSearchResults(results);
-    } catch {
-      if (id === searchIdRef.current) setSearchResults([]);
-    } finally {
-      if (id === searchIdRef.current) setSearchLoading(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    if (debounceRef.current) clearTimeout(debounceRef.current);
-    if (!searchQuery.trim()) {
-      setSearchResults([]);
-      return;
-    }
-    debounceRef.current = setTimeout(() => {
-      doSearch(searchQuery, searchMode);
-    }, DEBOUNCE_MS);
-
-    return () => {
-      if (debounceRef.current) clearTimeout(debounceRef.current);
-    };
-  }, [searchQuery, searchMode, doSearch]);
-
-  const handleSearchChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setSearchQuery(e.target.value);
-  };
-
-  const handleModeChange = (mode: SearchMode) => {
-    setSearchMode(mode);
-    setSearchQuery('');
-    setSearchResults([]);
-  };
-
   if (!isAuthenticated) {
     return <LoginRequiredView />;
   }
 
-  const isSearching = searchQuery.trim().length > 0;
   const tabLabels: TabKey[] = ['친구', '받은 요청', '보낸 요청'];
   const tabCounts: Record<TabKey, number> = {
     친구: friends.length,
